+++ src/app/import/services/import-file-emails.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed } from "@angular/core/testing";
+import { ImportFileEmailsService } from "./import-file-emails.service";
+
+describe("ImportFileEmailsService", () => {
+  let service: ImportFileEmailsService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ImportFileEmailsService);
+  });
+
+  it("should be created", () => {
+    expect(service).toBeTruthy();
+  });
+});
