--- src/app/import/services/import-parse-subscription-details.service.spec.ts
@@ -1,20 +1,0 @@
-import { TestBed } from "@angular/core/testing";
-import { buildTestModuleMetadata } from "../../../spec-helpers";
-import { ImportParseSubscriptionDetailsService } from "./import-parse-subscription-details.service";
-
-describe("ImportParseSubscriptionDetailsService", () => {
-  let service: ImportParseSubscriptionDetailsService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule(
-      buildTestModuleMetadata({
-        providers: [ImportParseSubscriptionDetailsService],
-      }),
-    );
-    service = TestBed.inject(ImportParseSubscriptionDetailsService);
-  });
-
-  it("should be created", () => {
-    expect(service).toBeTruthy();
-  });
-});
